+++ tests/Middleware/ExceptionResponderTest.php
@@ -5,13 +5,17 @@ declare(strict_types=1);
 namespace Yiisoft\ErrorHandler\Tests\Middleware;
 
 use DomainException;
+use HttpSoft\Message\Response;
 use HttpSoft\Message\ResponseFactory;
+use HttpSoft\Message\ResponseTrait;
 use HttpSoft\Message\ServerRequestFactory;
 use InvalidArgumentException;
+use LogicException;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\ResponseFactoryInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Message\StreamInterface;
 use Psr\Http\Server\RequestHandlerInterface;
 use Yiisoft\ErrorHandler\Middleware\ExceptionResponder;
 use Yiisoft\Http\Method;
@@ -61,6 +65,45 @@ final class ExceptionResponderTest extends TestCase
         $this->process($middleware);
     }
 
+    public function testPerformGetBodyOnResponse(): void
+    {
+        $middleware = $this->createMiddleware(performGetBodyOnResponse: true);
+        $request = (new ServerRequestFactory())->createServerRequest(Method::GET, 'http://example.com');
+        $handler = new class () implements RequestHandlerInterface {
+            public function handle(ServerRequestInterface $request): ResponseInterface
+            {
+                return new class () implements ResponseInterface {
+                    use ResponseTrait;
+
+                    public function getBody(): StreamInterface
+                    {
+                        throw new LogicException('test');
+                    }
+                };
+            }
+        };
+
+        $this->expectException(LogicException::class);
+        $this->expectExceptionMessage('test');
+        $middleware->process($request, $handler);
+    }
+
+    public function testSuccess(): void
+    {
+        $middleware = $this->createMiddleware(performGetBodyOnResponse: true);
+        $request = (new ServerRequestFactory())->createServerRequest(Method::GET, 'http://example.com');
+        $handler = new class () implements RequestHandlerInterface {
+            public function handle(ServerRequestInterface $request): ResponseInterface
+            {
+                return new Response();
+            }
+        };
+
+        $response = $middleware->process($request, $handler);
+
+        $this->assertSame(200, $response->getStatusCode());
+    }
+
     private function process(ExceptionResponder $middleware): ResponseInterface
     {
         return $middleware->process(
@@ -74,7 +117,10 @@ final class ExceptionResponderTest extends TestCase
         );
     }
 
-    private function createMiddleware(array $exceptionMap): ExceptionResponder
+    private function createMiddleware(
+        array $exceptionMap = [],
+        bool $performGetBodyOnResponse = false,
+    ): ExceptionResponder
     {
         return new ExceptionResponder(
             $exceptionMap,
@@ -84,6 +130,7 @@ final class ExceptionResponderTest extends TestCase
                     ResponseFactoryInterface::class => new ResponseFactory(),
                 ]),
             ),
+            $performGetBodyOnResponse,
         );
     }
 }